move request hooks and on error hooks to new folder. move on error hooks to its own class
