+++ Fabric.Authorization.IntegrationTests/Modules/ClientTests.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Security.Claims;
 using Fabric.Authorization.API;
 using Fabric.Authorization.API.Constants;
+using Fabric.Authorization.API.Infrastructure.PipelineHooks;
 using Fabric.Authorization.API.Models;
 using Fabric.Authorization.API.Modules;
 using Fabric.Authorization.Domain.Stores;
