+++ Fabric.Authorization.IntegrationTests/Modules/PermissionsTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Security.Claims;
 using Fabric.Authorization.API;
 using Fabric.Authorization.API.Constants;
+using Fabric.Authorization.API.Infrastructure.PipelineHooks;
 using Fabric.Authorization.API.Modules;
 using Fabric.Authorization.Domain.Stores;
 using Fabric.Authorization.Domain.Stores.CouchDB;
