+++ Fabric.Authorization.UnitTests/ModuleTestsBase.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Security.Claims;
 using Fabric.Authorization.API;
+using Fabric.Authorization.API.Infrastructure.PipelineHooks;
 using Fabric.Authorization.Domain.Models;
 using Fabric.Authorization.Domain.Stores;
 using Fabric.Authorization.UnitTests.Mocks;
